+++ lib/manifest.ps1
@@ -558,6 +558,8 @@ function default_architecture {
     $arch = get_config 'default-architecture'
     $system = if ([System.IntPtr]::Size -eq 8) { '64bit' } else { '32bit' }
 
+    if (Test-IsArmArchitecture) { $arch = 'arm' + ($system -replace 'bit') }
+
     if ($null -eq $arch) {
         $arch = $system
     } else {